requirements.txt
github3.py==4.0.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0